Remove JSON optimizations (fix #10)
There are some JSON optimizations in QuickStart that turned out to
cause problems in some cases. Since the optimizations are not
really needed, we might as well just remove them and treat the
require statements of JSON files the same as other files.

The optimizations would previously do the following conversion
at build-time:

file.json:

```json
{"name": "Foo"}
```

index.js:

```js
var name = require('./file.json').name;
// -->
var name = 'Foo';
```

That meant that if nothing else was used from the JSON file, the
JSON file itself would never be included in the built file.
However, this is causing issues when the accessed property is
actually supposed to refer to the evaluated result of the JSON
file, as seen in this example:

```js
// assume a JSON file of ['Foo', 'Bar']
var names = require('./names').join(', ');
```